+++ object/store/dual/reader.go
@@ -24,17 +24,17 @@ func (dual *Dual) ReadReaderFull(id id.ObjectID) (io.ReadCloser, error) {
 
 // ReadReaderContent reads an object's type, declared content length,
 // and content stream from the combined view.
-func (dual *Dual) ReadReaderContent(id id.ObjectID) (typ.Type, uint64, io.ReadCloser, error) {
+func (dual *Dual) ReadReaderContent(id id.ObjectID) (typ.Type, int, io.ReadCloser, error) {
 	return dual.reader.ReadReaderContent(id) //nolint:wrapcheck
 }
 
 // ReadSize reads an object's declared content length from the combined view.
-func (dual *Dual) ReadSize(id id.ObjectID) (uint64, error) {
+func (dual *Dual) ReadSize(id id.ObjectID) (int, error) {
 	return dual.reader.ReadSize(id) //nolint:wrapcheck
 }
 
 // ReadHeader reads an object's type and declared content length from the combined view.
-func (dual *Dual) ReadHeader(id id.ObjectID) (typ.Type, uint64, error) {
+func (dual *Dual) ReadHeader(id id.ObjectID) (typ.Type, int, error) {
 	return dual.reader.ReadHeader(id) //nolint:wrapcheck
 }
